requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
duckdb==0.10.0
python-dotenv==1.0.1 